+++ packages/patternfly-4/react-charts/src/declarations.d.ts
@@ -2,4 +2,5 @@ declare module 'victory-core' {
   export const Data: any;
   export const Helpers: any;
   export const Path: any;
+  export const TextSize: any;
 }