+++ packages/patternfly-4/react-core/src/components/Dropdown/DropdownGroup.d.ts
@@ -1,8 +1,9 @@
 import { FunctionComponent, HTMLProps, ReactNode } from 'react';
+import { Omit } from '../../helpers/typeUtils';
 
-export interface DropdownGroupProps extends HTMLProps<HTMLDivElement> {
+export interface DropdownGroupProps extends Omit<React.HTMLProps<HTMLDivElement>, 'label'> {
   children?: ReactNode;
-  label?: string;
+  label?: ReactNode;
 }
 
 declare const DropdownGroup: FunctionComponent<DropdownGroupProps>;