+++ packages/patternfly-4/react-core/src/components/Dropdown/dropdownConstants.d.ts
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { any } from 'glamor';
 
 export enum DropdownPosition {
   right = 'right',
@@ -16,7 +15,14 @@ export const DropdownContext: React.Context<{
   menuClass?: string,
   itemClass?: string,
   toggleClass?: string,
-  baseClass?: string
+  baseClass?: string,
+  baseComponent?: string,
+  sectionClass?: string,
+  sectionTitleClass?: string,
+  sectionComponent?: string,
+  disabledClass?: string,
+  hoverClass?: string,
+  separatorClass?: string
 }>
 
 export const DropdownArrowContext: React.Context<{