+++ packages/patternfly-4/react-core/src/components/Popover/Popover.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import Tippy from '@tippy.js/react';
-import { Instance as TippyInstance, BasicPlacement } from 'tippy.js';
+import { Instance as TippyInstance } from 'tippy.js';
 import { KEY_CODES } from '../../helpers/constants';
 import styles from '@patternfly/react-styles/css/components/Popover/popover';
 import { css, getModifier } from '@patternfly/react-styles';