+++ packages/patternfly-4/react-core/src/declarations.d.ts
@@ -3,4 +3,5 @@ declare module 'victory-core' {
   export const Data: any;
   export const Helpers: any;
   export const Path: any;
+  export const TextSize: any;
 }