+++ packages/patternfly-4/react-virtualized-extension/src/components/Virtualized/VirtualGrid.ts
@@ -29,7 +29,6 @@ import {
   CellCache,
   StyleCache
 } from './types';
-import { overflow } from 'styled-system';
 /**
  * Specifies the number of milliseconds during which to disable pointer events while a scroll is in progress.
  * This improves performance and makes scrolling smoother.
@@ -53,7 +52,7 @@ type ScrollPosition = {
 };
 
 type Props = {
-  'aria-label': string,
+  'aria-label'?: string,
   'aria-readonly'?: boolean,
 
   /**