+++ packages/patternfly-4/react-integration/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.24.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-integration@1.23.0...@patternfly/react-integration@1.24.0) (2019-06-19)
+
+
+### Features
+
+* **FormSelect:** Convert form select to typescript ([#2204](https://github.com/patternfly/patternfly-react/issues/2204)) ([373e012](https://github.com/patternfly/patternfly-react/commit/373e012))
+
+
+
+
+
 # [1.23.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-integration@1.21.0...@patternfly/react-integration@1.23.0) (2019-06-18)